requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
black==24.4.2
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.0
distro==1.9.0
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Migrate==4.0.5
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
msgspec==0.18.6
mypy-extensions==1.0.0
openai==1.35.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.0.0
pytest-cov==5.0.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.9.0
urllib3==2.2.2
Werkzeug==3.0.1
